#119 Moved common methods to seperate file
